requirements.txt
# Core Framework
Flask==2.3.3
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3

# ML Libraries
scikit-learn==1.3.0
imbalanced-learn==0.11.0
joblib==1.3.2

# Data Processing
requests==2.31.0
python-dotenv==1.0.0

# Configuration Management
PyYAML==6.0.1

# AWS SDK (for EC2 management)
boto3==1.34.0

# API Framework
fastapi==0.104.1
uvicorn==0.24.0

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.1.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Email for Alerts
smtplib-ssl==1.0.4

# Development and Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0